+++ resources/js/components/student/ShowStudents.vue
@@ -537,111 +537,9 @@
         <iframe :src="diceUrl" width="100%" style="height: 100vh"></iframe>
       </div>
     </b-modal>
-    <b-modal
-      :active.sync="isCardModalActive"
-      has-modal-card
-      full-screen
-      :can-cancel="false"
-    >
-      <div class="modal-card" style="width: auto">
-        <header class="modal-card-head">
-          <p class="modal-card-title">Random card</p>
-        </header>
-        <section class="modal-card-body is-relative" id="confetti-bg">
-          <img
-            id="deck"
-            src="/img/deck.png"
-            class="deck"
-            v-if="!showCard"
-            @click="revealCard"
-          />
-          <show-card
-            :card="randomCard"
-            class="deck"
-            :admin="false"
-            v-if="showCard"
-            :code="classroom.code"
-          ></show-card>
-        </section>
-        <footer
-          class="modal-card-foot columns is-multiline"
-          style="overflow-x: auto"
-        >
-          <div class="column is-narrow is-12-mobile is-flex has-all-centered">
-            <button
-              class="button"
-              type="button"
-              @click="isCardModalActive = false"
-            >
-              {{ trans.get("general.close") }}
-            </button>
-            <button
-              class="button is-link"
-              @click="
-                getRandomCard();
-                showCard = false;
-              "
-            >
-              <i class="far fa-redo-alt"></i>
-            </button>
-          </div>
-          <div
-            v-if="showCard"
-            class="column is-narrow is-flex has-all-centered"
-          >
-            <div class="field m-0">
-              <p class="control has-icons-left">
-                <span class="select">
-                  <select v-model="studentSelected">
-                    <option value="0">Student</option>
-                    <option
-                      :value="student.id"
-                      v-for="student in allStudents"
-                      :key="student.id"
-                    >
-                      {{ student.name }}
-                    </option>
-                  </select>
-                </span>
-                <span class="icon is-small is-left">
-                  <i class="fas fa-user"></i>
-                </span>
-              </p>
-            </div>
-            <button class="button is-primary" @click="assignCard('student')">
-              <i class="fas fa-user"></i>
-            </button>
-          </div>
-          <div
-            v-if="showCard"
-            class="column is-narrow is-flex has-all-centered"
-          >
-            <div class="field m-0">
-              <p class="control has-icons-left">
-                <span class="select">
-                  <select v-model="groupSelected">
-                    <option value="0">Groups</option>
-                    <option
-                      v-for="group in classroom.grouping[0].groups"
-                      :value="group.id"
-                      :key="group.id"
-                    >
-                      {{ group.name }}
-                    </option>
-                  </select>
-                </span>
-                <span class="icon is-small is-left">
-                  <i class="fas fa-user"></i>
-                </span>
-              </p>
-            </div>
-            <button class="button is-primary" @click="assignCard('group')">
-              <i class="fas fa-users"></i>
-            </button>
-          </div>
-        </footer>
-      </div>
-    </b-modal>
+    
+
+
     <b-modal
       :active.sync="isCountDownModalActive"
       has-modal-card
@@ -704,13 +602,14 @@
     </Impostor>
     <Videochat v-if="isVideoChatActive" :groups="groups" :code="classroom.code">
     </Videochat>
+    <random-card :card="randomCard" :admin="1" :code="classroom.code"></random-card>
   </div>
 </template>
 
 <script>
-import confetti from "canvas-confetti";
 import Impostor from "../utils/Impostor.vue";
 import Videochat from "../utils/Videochat.vue";
+import RandomCard from "../utils/RandomCard.vue";
 import InfiniteLoading from "vue-infinite-loading";
 
 export default {
@@ -745,6 +644,7 @@ export default {
       viewGrid: "",
       buttons: "",
       numItems: 5,
+      randomCard: null,
       isImpostorActive: false,
       isVideoChatActive: false,
       isQrModalActive: false,
@@ -761,10 +661,6 @@ export default {
       shuffledStudents: null,
       shuffledGroups: null,
       diceUrl: "",
-      randomCard: null,
-      showCard: false,
-      studentSelected: 0,
-      groupSelected: 0,
       event: null,
       page: 0,
       loading: false,
@@ -827,84 +723,6 @@ export default {
     refresh() {
       location.reload();
     },
-    assignCard(to) {
-      let card = this.randomCard.id;
-      let target;
-      if (to == "student") {
-        if (this.studentSelected == 0) {
-          this.$buefy.dialog.alert({
-            title: "Error",
-            message: "Please, select a student",
-            type: "is-danger",
-            hasIcon: true,
-            icon: "times-circle",
-            iconPack: "fa",
-            ariaRole: "alertdialog",
-            ariaModal: true,
-          });
-          return false;
-        } else {
-          target = this.studentSelected;
-        }
-      } else {
-        if (this.groupSelected == 0) {
-          this.$buefy.dialog.alert({
-            title: "Error",
-            message: "Please select a group",
-            type: "is-danger",
-            hasIcon: true,
-            icon: "times-circle",
-            iconPack: "fa",
-            ariaRole: "alertdialog",
-            ariaModal: true,
-          });
-          return false;
-        } else {
-          target = this.groupSelected;
-        }
-      }
-      axios
-        .post("/classroom/" + this.classroom.code + "/card/assign", {
-          type: to,
-          id: target,
-          card: card,
-        })
-        .then((response) => {
-          this.getRandomCard();
-          this.showCard = false;
-        });
-    },
-    revealCard() {
-      var audio = new Audio("/sound/victory.mp3");
-      audio.play();
-      setTimeout(() => {
-        var end = Date.now() + 15 * 125;
-
-        var colors = ["#bb0000", "#00bb00", "#0000bb", "#bbbb00"];
-
-        (function frame() {
-          confetti({
-            particleCount: 4,
-            angle: 60,
-            spread: 105,
-            origin: { x: 0 },
-            colors: colors,
-          });
-          confetti({
-            particleCount: 4,
-            angle: 120,
-            spread: 105,
-            origin: { x: 1 },
-            colors: colors,
-          });
-
-          if (Date.now() < end) {
-            requestAnimationFrame(frame);
-          }
-        })();
-        this.showCard = true;
-      }, 300);
-    },
     rollTheDice() {
       var audio = new Audio("/sound/dice.mp3");
       audio.play();
@@ -937,6 +755,7 @@ export default {
         .then((response) => {
           this.randomCard = response.data;
           this.isCardModalActive = true;
+          this.$forceUpdate();
         });
     },
     showClassCode() {
@@ -992,6 +811,7 @@ export default {
     Impostor,
     Videochat,
     InfiniteLoading,
+    RandomCard,
   },
   computed: {
     filteredDataObj() {