+++ resources/lang/ca/success_error.php
@@ -42,4 +42,6 @@ return [
 
     '403time' => 'Aquest repte estarà actiu el ',
     '403reqChallenge' => 'Aquest repte té un requeriment: ',
+    
+    'max_cards_reached' => 'Màxim de cartes superat!',
 ];